Used displayed flag for WHAM checking
When the WHAM (WEA Handset Action Message) message arrives, we
used to check all non-broadcasted messages for location recheck.
This caused a problem that a message could have been broadcasted
from cell broadcast service, but not shown by cell broadcast
receiver due to user's language setting mismatch.

Carriers use this language trick to test WHAM message since they
can't really move the device around in the lab. They insisted this
follows the ATIS spec, but totally denied this does not really
test the original purpose of WHAM.

The originaly purpose of WHAM is for device to re-check the
location again for those un-shown messages due to out of polygon,
not due to *language mismatch*.

Instead of using broadcasted flag for duplicte check, now use
the new flag we added in Android R, the displayed flag. This flag
indicates whether the emergency message was actually shown to the
user or not. Broadcasted flag only means cell broadcast service
broadcasted this message to cell broadcast receiver. It does not
guarantee the message was shown.

This should be enough to pass those WHAM tests.

Bug: 143229383
Test: Cell broadcast unit tests
Change-Id: I1fe409fb162d7ca30f0e1c475c4ca087e413bd37